types: Refactor types package * Remove contract types * Separate id and row_id
